remove util and move code to utils
