+++ integrations/nvidia/src/haystack_integrations/components/embedders/nvidia/text_embedder.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, List, Optional, Union
 
 from haystack import component, default_from_dict, default_to_dict
 from haystack.utils import Secret, deserialize_secrets_inplace
-from haystack_integrations.util.nvidia import NimBackend, is_hosted, url_validation
+from haystack_integrations.utils.nvidia import NimBackend, is_hosted, url_validation
 
 from .truncate import EmbeddingTruncateMode
