--- integrations/nvidia/src/haystack_integrations/util/nvidia/__init__.py
@@ -1,4 +1,0 @@
-from .nim_backend import Model, NimBackend
-from .util import is_hosted
-
-__all__ = ["NimBackend", "Model", "is_hosted"]
