+++ integrations/nvidia/src/haystack_integrations/utils/nvidia/__init__.py
@@ -1,3 +1,4 @@
-from .utils import url_validation
+from .nim_backend import Model, NimBackend
+from .utils import is_hosted, url_validation
 
-__all__ = ["url_validation"]
+__all__ = ["NimBackend", "Model", "is_hosted", "url_validation"]
