+++ integrations/nvidia/src/haystack_integrations/utils/nvidia/nim_backend.py
